file

+++ 0x01-variables_if_else_while/100-print_comb3.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	float num = 0;
+	int num = 0;
 
 	while (num < 90)
 	{
